added create note button to empty state todo list
